fix code complexity for sonar scan
