+++ lib/sdm.js
@@ -1,20 +1,21 @@
 const cds = require("@sap/cds/lib");
 const {
   getFolderIdByPath,
-  createFolder,
   createAttachment,
   deleteAttachmentsOfFolder,
   deleteFolderWithAttachments,
   readAttachment,
   renameAttachment
 } = require("../lib/handler");
-const { fetchAccessToken } = require("./util/index");
+const {
+  fetchAccessToken,
+  checkAttachmentsToRename,
+  getParentId 
+} = require("./util/index");
 const {
   getDraftAttachments,
   getURLsToDeleteFromAttachments,
   getURLFromAttachments,
-  getFolderIdForEntity,
-  getExistingAttachments
 } = require("../lib/persistence");
 const { duplicateDraftFileErr, emptyFileErr } = require("./util/messageConsts");
 
@@ -49,10 +50,7 @@ module.exports = class SDMAttachmentsService extends (
     const attachment_val = await getDraftAttachments(attachments, req);
 
     if (attachment_val.length > 0) {
-      const duplicateDraftFilesErrMsg = this.isFileNameDuplicateInDrafts(attachment_val);
-      if (duplicateDraftFilesErrMsg != "") {
-        req.reject(409, duplicateDraftFileErr(duplicateDraftFilesErrMsg));
-      }
+      await this.isFileNameDuplicateInDrafts(attachment_val);
 
       const token = await fetchAccessToken(
         this.creds,
@@ -61,75 +59,54 @@ module.exports = class SDMAttachmentsService extends (
       const attachment_val_rename = attachment_val.filter(attachment => attachment.HasActiveEntity === true);
       const attachment_val_create = attachment_val.filter(attachment => attachment.HasActiveEntity === false);
       const attachmentIDs = attachment_val_rename.map(attachment => attachment.ID);
-       const modifiedAttachments = [];
-      if(attachment_val_rename.length>0){
-      const matchedAttachments = await getExistingAttachments(attachmentIDs, attachments);
-      attachment_val_rename.forEach(draftAttachment => {
-        const correspondingAttachment = matchedAttachments.find(attachment => attachment.ID === draftAttachment.ID);
-        if(correspondingAttachment && correspondingAttachment.filename !== draftAttachment.filename) {
-          modifiedAttachments.push({ID:draftAttachment.ID, url: draftAttachment.url, name: draftAttachment.filename, prevname: correspondingAttachment.filename,folderId:correspondingAttachment.folderId});
-        }
-      });
-      }
-      if (modifiedAttachments.length>0) {
-        const failedReq = await this.onRename(
-          modifiedAttachments,
-          this.creds,
-          token,
-          req
-        );
-        let errorResponse = "";
-        failedReq.forEach((attachment) => {
-            attachment = attachment.replace("Child", "Attachment with name ");
-            errorResponse = errorResponse + "\n" + attachment;
-        });
-        if (errorResponse !== "") {
-            req.info(200, errorResponse);
-        }
+      const modifiedAttachments = [];
+
+      modifiedAttachments = await checkAttachmentsToRename(attachment_val_rename, attachmentIDs, attachments)
+
+      if(modifiedAttachments.length>0){
+        await this.rename(modifiedAttachments, token, req)
       }
       if(attachment_val_create.length>0){
-      const folderIds = await getFolderIdForEntity(attachments, req);
-      let parentId = "";
-      if (folderIds?.length == 0) {
-        const folderId = await getFolderIdByPath(
-          req,
-          this.creds,
-          token,
-          attachments
-        );
-        if (folderId) {
-          parentId = folderId;
-        } else {
-          const response = await createFolder(
-            req,
-            this.creds,
-            token,
-            attachments
-          );
-          parentId = response.data.succinctProperties["cmis:objectId"];
-        }
-      } else {
-        parentId = folderIds ? folderIds[0].folderId : "";
-      }
-      const failedReq = await this.onCreate(
-        attachment_val_create,
-        this.creds,
-        token,
-        attachments,
-        req,
-        parentId
-      );
-      let errorResponse = "";
-      failedReq.forEach((attachment) => {
-        attachment = attachment.replace("Child", "Attachment with name ");
-        errorResponse = errorResponse + "\n" + attachment;
-      });
-      if (errorResponse != "") req.info(200, errorResponse);
+        await this.create(attachment_val_create, attachments, req, token)
       }
     }
   }
+  async rename(modifiedAttachments, token, req){
+    const failedReq = await this.onRename(
+      modifiedAttachments,
+      this.creds,
+      token,
+      req
+    );
+    let errorResponse = "";
+    failedReq.forEach((attachment) => {
+        attachment = attachment.replace("Child", "Attachment with name ");
+        errorResponse = errorResponse + "\n" + attachment;
+    });
+    if (errorResponse !== "") {
+        req.info(200, errorResponse);
+    }
+  }
 
-  isFileNameDuplicateInDrafts(data) {
+  async create(attachment_val_create, attachments, req, token){
+    let parentId = getParentId(attachments,req,token)
+    const failedReq = await this.onCreate(
+      attachment_val_create,
+      this.creds,
+      token,
+      attachments,
+      req,
+      parentId
+    );
+    let errorResponse = "";
+    failedReq.forEach((attachment) => {
+      attachment = attachment.replace("Child", "Attachment with name ");
+      errorResponse = errorResponse + "\n" + attachment;
+    });
+    if (errorResponse != "") req.info(200, errorResponse);
+  }
+
+  async isFileNameDuplicateInDrafts(data) {
     let fileNames = [];
     for (let index in data) {
       fileNames.push(data[index].filename);
@@ -141,7 +118,9 @@ module.exports = class SDMAttachmentsService extends (
         })
       ),
     ];
-    return duplicates.join(", ");
+    if (duplicates.length != 0) {
+      req.reject(409, duplicateDraftFileErr(duplicates.join(", ")));
+    }
   }
 
   async attachDeletionData(req) {
