+++ lib/util/index.js
@@ -3,6 +3,17 @@ const cds = require("@sap/cds");
 const requests = xssec.requests;
 const NodeCache = require("node-cache");
 const cache = new NodeCache();
+
+const {
+  getFolderIdForEntity,
+  getExistingAttachments
+} = require("../lib/persistence");
+
+const {
+  getFolderIdByPath,
+  createFolder
+} = require("../lib/handler");
+
 async function fetchAccessToken(credentials, jwt) {
   let decoded_token_jwt = decodeAccessToken(jwt);
   let access_token = cache.get(decoded_token_jwt.email); // to check if token exists
@@ -55,7 +66,49 @@ function getConfigurations() {
   return cds.env.requires?.["sdm"]?.settings || {};
 }
 
+async function checkAttachmentsToRename(attachment_val_rename, attachmentIDs, attachments){
+  if(attachment_val_rename.length>0){
+    const matchedAttachments = await getExistingAttachments(attachmentIDs, attachments);
+    attachment_val_rename.forEach(draftAttachment => {
+      const correspondingAttachment = matchedAttachments.find(attachment => attachment.ID === draftAttachment.ID);
+      if(correspondingAttachment && correspondingAttachment.filename !== draftAttachment.filename) {
+        modifiedAttachments.push({ID:draftAttachment.ID, url: draftAttachment.url, name: draftAttachment.filename, prevname: correspondingAttachment.filename,folderId:correspondingAttachment.folderId});
+      }
+    });
+    }
+
+    return modifiedAttachments
+}
+
+async function getParentId(attachments,req,token){
+  const folderIds = await getFolderIdForEntity(attachments, req);
+  let parentId = "";
+  if (folderIds?.length == 0) {
+    const folderId = await getFolderIdByPath(
+      req,
+      this.creds,
+      token,
+      attachments
+    );
+    if (folderId) {
+      parentId = folderId;
+    } else {
+      const response = await createFolder(
+        req,
+        this.creds,
+        token,
+        attachments
+      );
+      parentId = response.data.succinctProperties["cmis:objectId"];
+    }
+  } else {
+    parentId = folderIds ? folderIds[0].folderId : "";
+  }
+}
+
 module.exports = {
   fetchAccessToken,
   getConfigurations,
+  checkAttachmentsToRename,
+  getParentId
 };
